Add 'preamble' to LB/DNS/TLS stories to present the relevent vocabulary
I have mixed feelings about this change, but I think it could be interesting to try it out. The idea is, prior to any work from the participant, give them the vocabulary and map of how the different concepts fit together. I'm hoping that it makes the rest of the work a bit more valauble.