Revise checklist to restrict LastPass access
We might have CF Onboarding Week participants who are interested in the
open-source stories, but are not authorized for closed-source content.
In cases like this, one option is to omit LastPass access. If the
participant needs access to work on an open-source story, facilitators
can recommend that they pair with someone who is authorized.